refactor: split automation_server.rs into directory with 5 files (#127 step 11)
Split 1914-line automation_server.rs into:
- mod.rs: start(), build_router(), auth_middleware(), ServerState
- types.rs: AutomationRequest/Response, body types, REGISTERED_ROUTES
- handlers_backend.rs: health, terminal_write/output, api_docs, terminals
- handlers_bridge.rs: 30+ frontend bridge handlers
- helpers.rs: bridge_request(), ok_json(), err_json()

389 tests passing.
